✨feat: prevent users from deleting questions from others

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import type { QuestionRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -11,10 +12,15 @@ export class DeleteQuestionUseCase {
   constructor(private questionRepository: QuestionRepository) {}
   async execute({
     questionId,
+    authorId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
     if (!question) {
       throw new Error('Question not found')
+    }
+    
+    if(authorId !== question.authorId.toString()){
+      throw new Error('You are not allower to delete this question')
     }
 
     await this.questionRepository.delete(question)
